# swagger.json
{
  "openapi": "3.0.0",
  "paths": {
    "/api/v1/order-histories": {
      "post": {
        "operationId": "OrderHistoriesController_create_v1",
        "parameters": [
          {
            "in": "header",
            "required": false,
            "name": "x-custom-lang",
            "schema": {
              "example": "en"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateOrderHistoryDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrderHistory"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": ["Orderhistories"]
      },
      "get": {
        "operationId": "OrderHistoriesController_findAll_v1",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "orderId",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "header",
            "required": false,
            "name": "x-custom-lang",
            "schema": {
              "example": "en"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InfinityPaginationOrderHistoryResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": ["Orderhistories"]
      }
    },
    "/api/v1/order-histories/{id}": {
      "get": {
        "operationId": "OrderHistoriesController_findById_v1",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "header",
            "required": false,
            "name": "x-custom-lang",
            "schema": {
              "example": "en"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrderHistory"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": ["Orderhistories"]
      }
    },
...
    "/api/v1/orders/{id}": {
      "get": {
        "operationId": "OrdersController_findById_v1",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "header",
            "required": false,
            "name": "x-custom-lang",
            "schema": {
              "example": "en"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Order"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": ["Orders"]
      },
      "delete": {
        "operationId": "OrdersController_remove_v1",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "header",
            "required": false,
            "name": "x-custom-lang",
            "schema": {
              "example": "en"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": ["Orders"]
      }
    },
...
  },
  "info": {
    "title": "API",
    "description": "API docs",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
      "UserDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "userId"
          }
        },
        "required": ["id"]
      },
      "OrderDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          }
        },
        "required": ["id"]
      },
      "OrderStatusEnum": {
        "type": "string",
        "enum": [
          "newPending",
          "confirmed",
          "packing",
          "shipped",
          "inTransit",
          "delivered",
          "cancelled",
          "returnRequested",
          "returned",
          "failedDelivery"
        ],
        "description": "Trạng thái đơn hàng trong lịch sử thay đổi"
      },
      "CreateOrderHistoryDto": {
        "type": "object",
        "properties": {
          "changeByUser": {
            "$ref": "#/components/schemas/UserDto"
          },
          "order": {
            "$ref": "#/components/schemas/OrderDto"
          },
          "status": {
            "description": "Trạng thái đơn hàng trong lịch sử thay đổi",
            "example": "newPending",
            "allOf": [
              {
                "$ref": "#/components/schemas/OrderStatusEnum"
              }
            ]
          }
        },
        "required": ["changeByUser", "order", "status"]
      },
      "FileType": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "cbcfa8b8-3a25-4adb-a9c6-e325f0d0f3ae"
          },
          "path": {
            "type": "string",
            "example": "https://example.com/path/to/file.jpg"
          }
        },
        "required": ["id", "path"]
      },
      "Role": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string",
            "example": "admin"
          }
        },
        "required": ["id", "name"]
      },
      "User": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "email": {
            "type": "string",
            "example": "john.doe@example.com"
          },
          "firstName": {
            "type": "string",
            "example": "email"
          },
          "lastName": {
            "type": "string",
            "example": "Doe"
          },
          "photo": {
            "$ref": "#/components/schemas/FileType"
          },
          "role": {
            "$ref": "#/components/schemas/Role"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string"
          },
          "deletedAt": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "id",
          "email",
          "firstName",
          "lastName",
          "photo",
          "role",
          "createdAt",
          "updatedAt",
          "deletedAt"
        ]
      },
      "Customer": {
        "type": "object",
        "properties": {
          "note": {
            "type": "string",
            "nullable": true
          },
          "address": {
            "type": "string",
            "nullable": false
          },
          "email": {
            "type": "string",
            "nullable": true
          },
          "phone": {
            "type": "string",
            "nullable": false
          },
          "name": {
            "type": "string",
            "nullable": false
          },
          "id": {
            "type": "string"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "note",
          "address",
          "email",
          "phone",
          "name",
          "id",
          "createdAt",
          "updatedAt"
        ]
      },
      "OrderItem": {
        "type": "object",
        "properties": {
          "productName": {
            "type": "string",
            "example": "World smallest violin"
          },
          "quantity": {
            "type": "number",
            "example": 2
          },
          "unitPrice": {
            "type": "number",
            "example": 50000
          }
        },
        "required": ["productName", "quantity", "unitPrice"]
      },
      "Order": {
        "type": "object",
        "properties": {
          "customer": {
            "nullable": false,
            "allOf": [
              {
                "$ref": "#/components/schemas/Customer"
              }
            ]
          },
          "item": {
            "nullable": false,
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/OrderItem"
            }
          },
          "totalAmount": {
            "type": "number",
            "nullable": false
          },
          "status": {
            "nullable": false,
            "description": "Trạng thái đơn hàng trong lịch sử thay đổi",
            "example": "newPending",
            "allOf": [
              {
                "$ref": "#/components/schemas/OrderStatusEnum"
              }
            ]
          },
          "deliveredDate": {
            "format": "date-time",
            "type": "string",
            "nullable": true
          },
          "expectedDeliveryDate": {
            "format": "date-time",
            "type": "string",
            "nullable": true
          },
          "orderDate": {
            "format": "date-time",
            "type": "string",
            "nullable": false
          },
          "deliveryAddress": {
            "type": "string",
            "nullable": false
          },
          "orderCode": {
            "type": "string",
            "nullable": false
          },
          "id": {
            "type": "string"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "customer",
          "item",
          "totalAmount",
          "status",
          "deliveredDate",
          "expectedDeliveryDate",
          "orderDate",
          "deliveryAddress",
          "orderCode",
          "id",
          "createdAt",
          "updatedAt"
        ]
      },
      "OrderHistory": {
        "type": "object",
        "properties": {
          "changeByUser": {
            "nullable": false,
            "allOf": [
              {
                "$ref": "#/components/schemas/User"
              }
            ]
          },
          "order": {
            "nullable": false,
            "allOf": [
              {
                "$ref": "#/components/schemas/Order"
              }
            ]
          },
          "status": {
            "nullable": false,
            "description": "Trạng thái đơn hàng trong lịch sử thay đổi",
            "example": "newPending",
            "allOf": [
              {
                "$ref": "#/components/schemas/OrderStatusEnum"
              }
            ]
          },
          "id": {
            "type": "string"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "changeByUser",
          "order",
          "status",
          "id",
          "createdAt",
          "updatedAt"
        ]
      },
      "InfinityPaginationOrderHistoryResponseDto": {
        "type": "object",
        "properties": {
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/OrderHistory"
            }
          },
          "hasNextPage": {
            "type": "boolean",
            "example": true
          }
        },
        "required": ["data", "hasNextPage"]
      },
...
    }
  }
}